feat(demo-player-angular-configs): default buttons
Buttons now provide possiblity to fill the textareas for configuration and css with default values
Added link to player-configs libs

Update README.md

feat(player-configs): added multiple buttons for default configs

added multiple buttons for default configs
