+++ libs/player-configs/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/configs.module';
 export * from './lib/Factory';
+export * as CONFIGS from './lib/configs';
 export * from './lib/elements/playerHTML';
 export * from './lib/elements/SimplePlaylist';
